moved demo image (easier to delete latyer)
